fastmcp/requirements.txt
fastmcp>=0.2.0
langchain-mcp-adapters>=0.2.1
langchain-core>=0.1.0
langchain>=0.1.0
python-dotenv>=1.0.0
httpx>=0.25.0